refactor: relocated IPluginService and split it in multiple interfaces
