+++ src/Adapters/Driven/Marketplace.Client/Marketplace.Client.csproj
@@ -9,4 +9,8 @@
         <PackageReference Include="Microsoft.Extensions.DependencyInjection"/>
         <PackageReference Include="Microsoft.Extensions.Http"/>
     </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\..\Application\Domain\Model\Model.csproj" />
+    </ItemGroup>
 </Project>
