+++ src/Application/Domain/Model/Plugin/MarketPlaceProjectType.cs
@@ -1,0 +1,7 @@
+namespace ScriptBee.Domain.Model.Plugin;
+
+public enum MarketPlaceProjectType
+{
+    Plugin,
+    Bundle,
+}
