+++ src/Application/Domain/Model/Plugin/PluginVersion.cs
@@ -1,3 +1,3 @@
-namespace ScriptBee.Marketplace.Client.Data;
+namespace ScriptBee.Domain.Model.Plugin;
 
 public record PluginVersion(string Url, Version Version, string ManifestUrl);
