+++ src/Application/Ports/Driving/UseCases.Plugin/IGetAvailablePluginsUseCase.cs
@@ -1,0 +1,10 @@
+using ScriptBee.Domain.Model.Plugin;
+
+namespace ScriptBee.UseCases.Plugin;
+
+public interface IGetAvailablePluginsUseCase
+{
+    Task<IEnumerable<MarketPlacePlugin>> GetMarketPlugins(
+        CancellationToken cancellationToken = default
+    );
+}
