+++ src/Application/Ports/Driving/UseCases.Plugin/IInstallPluginUseCase.cs
@@ -1,0 +1,10 @@
+namespace ScriptBee.UseCases.Plugin;
+
+public interface IInstallPluginUseCase
+{
+    Task InstallPlugin(
+        string pluginId,
+        string version,
+        CancellationToken cancellationToken = default
+    );
+}
