+++ src/Application/Ports/Driving/UseCases.Plugin/IUninstallPluginUseCase.cs
@@ -1,0 +1,6 @@
+﻿namespace ScriptBee.UseCases.Plugin;
+
+public interface IUninstallPluginUseCase
+{
+    void UninstallPlugin(string pluginId, string pluginVersion);
+}
